+++ src/connectors/items-by-id/my-list/items.derive.js
@@ -13,6 +13,10 @@ export function deriveMyListItems(response) {
    * @save_status {string} A string value (unsaved, saving, saved)
    */
   return response.map((item) => {
+    // Status of 2 means it has been deleted and shouldn't be stored
+    // Unclear why these items are being sent. This needs to be resolved by Backend
+    if (item?.status === '2') return
+
     return {
       item_id: item?.item_id,
       resolved_id: item?.resolved_id,